requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==65.0.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
flake8==7.1.1
gunicorn==23.0.0
idna==3.10
mccabe==0.7.0
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
